+++ src/main/java/com/seeu/ywq/trend/dvo/PublishVOVideo.java
@@ -3,14 +3,14 @@ package com.seeu.ywq.trend.dvo;
 import java.math.BigDecimal;
 
 public class PublishVOVideo extends PublishVO {
-    private BigDecimal unlockPrice;//解锁需要金额
+    private Long unlockPrice;//解锁需要金额
     private PublishVideoVO video;
 
-    public BigDecimal getUnlockPrice() {
+    public Long getUnlockPrice() {
         return unlockPrice;
     }
 
-    public void setUnlockPrice(BigDecimal unlockPrice) {
+    public void setUnlockPrice(Long unlockPrice) {
         this.unlockPrice = unlockPrice;
     }
